[cita chain] split extras
